version finale avec partie 2 comprise

+++ src/main/Multiplication.java
@@ -6,7 +6,7 @@ public class Multiplication extends Operation{
         super(op1, op2);
     }
 
-    public int valeur(){
+    public double valeur(){
         return getOPerande1().valeur()*getOPerande2().valeur();
     }
